Moved library FTS into it's own module.
