+++ windows/ReactNativePSPDFKit/ReactNativePSPDFKit/PSPDFKitModule.cs
@@ -8,17 +8,13 @@
 //
 
 using PSPDFKit;
-using PSPDFKit.Search;
-using PSPDFKitFoundation.Search;
 using ReactNative.Bridge;
 using System;
 using System.Collections.Generic;
-using System.Diagnostics;
 using System.Reflection;
 using System.Threading.Tasks;
 using Windows.Storage;
 using Windows.UI.Popups;
-using Newtonsoft.Json.Linq;
 
 namespace ReactNativePSPDFKit
 {
@@ -29,7 +25,6 @@ namespace ReactNativePSPDFKit
     {
         private readonly PDFViewPage _pdfViewPage;
         private string VERSION_KEY = "versionString";
-        private readonly Dictionary<string,Library> _libraries = new Dictionary<string, Library>();
 
         public PSPDFKitModule(ReactContext reactContext, PDFViewPage pdfViewPage) : base(reactContext)
         {
@@ -77,228 +72,10 @@ namespace ReactNativePSPDFKit
         }
 
         /// <summary>
-        /// Open a search library from a ms path.
-        /// Multiple libraries maybe open at once. Use the name to reference each library.
-        /// Promise will resolve with true if library is opened. Promise will reject if folder is inaccessible.
-        /// <param name="libraryName">Name to give the library</param>
-        /// <param name="uri">A path to a folder to index. The application must have permission to the path.</param>
-        /// See https://docs.microsoft.com/en-us/windows/uwp/files/file-access-permissions
-        /// </summary>
-        [ReactMethod]
-        public void OpenLibrary(string libraryName, string path, IPromise promise)
-        {
-            DispatcherHelpers.RunOnDispatcher(async () =>
-            {
-                try
-                {
-                    Sdk.Initialize(_pdfViewPage.Pdfview.License);
-
-                    // If we have already opened a library with the same name, reject.
-                    if (_libraries.ContainsKey(libraryName))
-                    {
-                        promise.Reject(new Exception($"{libraryName} has already been added."));
-                        return;
-                    }
-                    _libraries.Add(libraryName, await Library.OpenLibraryAsync(libraryName));
-
-                    var storageFolder = await StorageFolder.GetFolderFromPathAsync(path);
-                    
-                    // Queue up the PDFs in the folder for indexing.
-                    await _libraries[libraryName].EnqueueDocumentsInFolderAsync(storageFolder);
-
-                    promise.Resolve(true);
-                }
-                catch (Exception e)
-                {
-                    promise.Reject(e);
-                }
-            });
-        }
-
-        /// <summary>
-        /// Open a search library with the use of a folder picker.
-        /// Multiple libraries maybe open at once. Use the name to reference each library.
-        /// Promise will resolve with true if library is opened. Promise will reject if folder is inaccessible.
-        /// <param name="libraryName">Name to give the library</param>
-        /// </summary>
-        [ReactMethod]
-        public void OpenLibraryPicker(string libraryName, IPromise promise)
-        { 
-            DispatcherHelpers.RunOnDispatcher(async () =>
-            {
-                try
-                {
-                    Sdk.Initialize(_pdfViewPage.Pdfview.License);
-
-                    // If we have already opened a library with the same name, reject.
-                    if (_libraries.ContainsKey(libraryName))
-                    {
-                        promise.Reject(new Exception($"{libraryName} has already been added."));
-                        return;
-                    }
-                    _libraries.Add(libraryName, await Library.OpenLibraryAsync(libraryName));
-
-                    // Allow the user to choose a folder to index.
-                    var folderPicker = new Windows.Storage.Pickers.FolderPicker
-                    {
-                        SuggestedStartLocation = Windows.Storage.Pickers.PickerLocationId.Desktop
-                    };
-                    folderPicker.FileTypeFilter.Add("*");
-
-                    var folder = await folderPicker.PickSingleFolderAsync();
-                    if (folder != null)
-                    {
-                        // Queue up the PDFs in the folder for indexing.
-                        await _libraries[libraryName].EnqueueDocumentsInFolderAsync(folder);
-
-                        promise.Resolve(true);
-                    } else
-                    {
-                        promise.Reject(new System.IO.FileNotFoundException("Folder not accessible"));
-                    }
-                }
-                catch (Exception e)
-                {
-                    promise.Reject(e);
-                }
-            });
-        }
-
-        /// <summary>
-        /// Search a given library for the searchTerm and provide a promise
-        /// to retreive the page indexes of found instances.
-        /// 
-        /// Json return example
-        /// [
-        ///     {
-        ///         "uid": "{2B94FFD0-F846-4902-8A11-75C3D1E5B2A3}/default.pdf",
-        ///         "pageResults": [ 2 ]
-        ///     }
-        /// ]
-        /// or if previews are generated.
-        /// [
-        ///     {
-        ///         "uid": "{2B94FFD0-F846-4902-8A11-75C3D1E5B2A3}/default.pdf",
-        ///         "pageIndex": 2,
-        ///         "previewText": "example in PSPDFKit.",
-        ///         "rangeInText": {
-        ///             "position": 27,
-        ///             "length": 8
-        ///         },
-        ///         "rangeInPreviewText": {
-        ///             "position": 11,
-        ///             "length": 8
-        ///         },
-        ///         "annotationId": 55
-        ///     },
-        /// ]
-        /// 
-        /// </summary>
-        /// <param name="libraryName">Name of library to search.</param>
-        /// <param name="searchLibraryQuery">Configuration of query </param>
-        /// Example of configuration. All Items except search string are optional.
-        /// const libraryConfiguration = {
-        ///     searchString: "pspdfkit",
-        ///     excludeAnnotations: false,
-        ///     excludeDocumentText: false,
-        ///     matchExactPhrases: false,
-        ///     maximumSearchResultsPerDocument: 0,
-        ///     maximumSearchResultsTotal: 500,
-        ///     maximumPreviewResultsPerDocument: 0,
-        ///     maximumPreviewResultsTotal: 500,
-        ///     generateTextPreviews: true,
-        ///     previewRange: { position: 20, length: 120 }
-        /// }
-        [ReactMethod]
-        public void SearchLibrary(string libraryName, JObject searchLibraryQuery, IPromise promise)
-        {
-            DispatcherHelpers.RunOnDispatcher(async () =>
-            {
-                // Find the library to search and reject if not present.
-                if (!_libraries.TryGetValue(libraryName, out var library))
-                {
-                    promise.Resolve(new Exception($"Library {libraryName} not loaded. Please use OpenLibrary."));
-                    return;
-                }
-                await library.WaitForAllIndexingTasksToFinishAsync();
-
-                var libraryQuery = JsonUtils.ToLibraryQuery(searchLibraryQuery);
-
-                TaskCompletionSource<IList<LibraryPreviewResult>> previewsFromHandlerTcs = null;
-                TaskCompletionSource<IDictionary<string, LibraryQueryResult>> resultsFromHandlerTcs = null;
-                if (libraryQuery.GenerateTextPreviews)
-                {
-                    previewsFromHandlerTcs = GetPreviewCompleteTcs(library);
-                }
-                else
-                {
-                    resultsFromHandlerTcs = GetSearchCompleteTcs(library);
-                }
-
-                // We can do the searching in the background as the callbacks will receive the results.
-#pragma warning disable CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
-                library.SearchAsync(libraryQuery);
-#pragma warning restore CS4014 // Because this call is not awaited, execution of the current method continues before the call is completed
-
-                if (libraryQuery.GenerateTextPreviews)
-                {
-                    var previewResults = await previewsFromHandlerTcs.Task;
-                    promise.Resolve(JsonUtils.PreviewResultsToJson(previewResults));
-                }
-                else
-                {
-                    var resultsFromHandler = await resultsFromHandlerTcs.Task;
-                    promise.Resolve(JsonUtils.SearchResultsToJson(resultsFromHandler));
-                }
-            });
-        }
-
-        [ReactMethod]
-        public void DeleteAllLibraries()
-        {
-            DispatcherHelpers.RunOnDispatcher(async () =>
-            {
-                foreach (var library in _libraries.Values)
-                {
-                    await library.CancelAllTasksAsync();
-                }
-                await Library.DeleteAllLibrariesAsync(); });
-        }
-        
-
-        private static TaskCompletionSource<IDictionary<string, LibraryQueryResult>> GetSearchCompleteTcs(Library library)
-        {
-            var resultsFromHandlerTcs = new TaskCompletionSource<IDictionary<string, LibraryQueryResult>>();
-            void SearchHandler(object sender, IDictionary<string, LibraryQueryResult> args)
-            {
-                library.OnSearchComplete -= SearchHandler;
-                resultsFromHandlerTcs.SetResult(args);
-            }
-
-            library.OnSearchComplete += SearchHandler;
-
-            return resultsFromHandlerTcs;
-        }
-
-        private static TaskCompletionSource<IList<LibraryPreviewResult>> GetPreviewCompleteTcs(Library library)
-        {
-            var previewsFromHandlerTcs = new TaskCompletionSource<IList<LibraryPreviewResult>>();
-            void Previewhandler(object sender, IList<LibraryPreviewResult> args)
-            {
-                library.OnSearchPreviewComplete -= Previewhandler;
-                previewsFromHandlerTcs.SetResult(args);
-            }
-
-            library.OnSearchPreviewComplete += Previewhandler;
-
-            return previewsFromHandlerTcs;
-        }
-
-        /// <summary>
         /// Opens the native file picker.
         /// </summary>
         /// <returns>The file chosen in the file picker.</returns>
-        private async Task<Windows.Storage.StorageFile> PickPDF()
+        private static async Task<Windows.Storage.StorageFile> PickPDF()
         {
             var picker = new Windows.Storage.Pickers.FileOpenPicker
             {
