+++ windows/ReactNativePSPDFKit/ReactNativePSPDFKit/PSPDFKitPackage.cs
@@ -36,6 +36,7 @@ namespace ReactNativePSPDFKit
             return new List<INativeModule>
             {
                 new PSPDFKitModule(reactContext, _pspdfkitViewManger.PdfViewPage),
+                new LibraryModule(reactContext, _pspdfkitViewManger.PdfViewPage.Pdfview.License)
             };
         }
