+++ windows/ReactNativePSPDFKit/ReactNativePSPDFKit/ReactNativePSPDFKit.csproj
@@ -111,6 +111,7 @@
     <Compile Include="Events\PdfViewDataReturnedEvent.cs" />
     <Compile Include="Events\PdfViewDocumentSaveFailedEvent.cs" />
     <Compile Include="JsonUtils.cs" />
+    <Compile Include="LibraryModule.cs" />
     <Compile Include="PSPDFKitModule.cs" />
     <Compile Include="PDFViewPage.xaml.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
@@ -136,7 +137,7 @@
     <SDKReference Include="Microsoft.VCLibs, Version=14.0">
       <Name>Visual C++ 2015 Runtime for Universal Windows Platform Apps</Name>
     </SDKReference>
-    <SDKReference Include="PSPDFKitSDK, Version=1.6.0">
+    <SDKReference Include="PSPDFKitSDK, Version=1.6.1">
       <Name>PSPDFKit for UWP</Name>
     </SDKReference>
   </ItemGroup>
